requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
flask==3.0.0
psycopg2-binary==2.9.9
